Move client into frontend
