+++ frontend/public/favicon-white.png
