+++ frontend/src/Dashboard/CategoryChips.tsx
