+++ frontend/src/Dashboard/Recipes/RecipeCard.tsx
@@ -5,7 +5,7 @@ import CardMedia from '@material-ui/core/CardMedia';
 import Chip from '@material-ui/core/Chip';
 import { Typography } from '@material-ui/core';
 import { useHistory } from 'react-router-dom';
-import { IRecipeFields } from '../../utils/contentful/fetchData';
+import { IRecipeFields } from '../../utils/contentful/types';
 import { LanguageContext } from '../../Language/LanguageProvider';
 import { dietTypes } from '../Filters/FiltersDietType';
 import { mealTypes } from '../Filters/FiltersMealType';
