+++ frontend/src/shared/DividerBlue.tsx
